refactor(sports): split middleware to features
TH-1377

The Sports app had a different setup for middleware compared to the
Events and the Hobbies apps, because it also had a dynamic short URL
redirection feature on. This feature allows CMS users to create short
URLs from CMS redirection table.

This refactoring will split the middleware into separate features for
better organization and maintainability. The target is to move those
features in a shared package, from where also the other apps can easily
activate those features. Also, this will clean up the middleware
codebase and make it much more readable and maintainable.
