+++ apps/sports-helsinki/src/edge-runtime-compatible/middlewares/shortUrls/dynamicUrls/index.ts
@@ -1,0 +1,1 @@
+export * from './redirectsCache';
