+++ apps/sports-helsinki/src/edge-runtime-compatible/middlewares/shortUrls/dynamicUrls/types.ts
@@ -1,0 +1,2 @@
+export type RedirectAddress = string;
+export type Redirects = Record<RedirectAddress, RedirectAddress>;
