+++ apps/sports-helsinki/src/edge-runtime-compatible/middlewares/shortUrls/index.ts
@@ -1,0 +1,2 @@
+export * from './redirectHandler';
+export * from './dynamicUrls';
